Una especie de patrón Facade se aplica en unqfy, teniendo ahora servicios a los cuáles delegarle las consultas
